Added CheckRectDraw and CustomRect, Updated RectDraw
CustomRect(previously CRect) is now a separate program that is used by both RectDraw and CheckRectDraw. 
It is a custom implementation of the open CV Rect
Both RectDraw and CheckRectDraw implement the same techniques to detect the bubbles and draw the rectangles. 
The difference is that CheckRectDraw draws each rectangle one-by-one whereas RectDraw draws them all at once. 
CHECKRECTDRAW was used to check if the rectangles were being drawn in reading order(from left to right)